# src/main/java/io/hexlet/flightbooking/dto/PassengerDto.java
package io.hexlet.flightbooking.dto;

import java.time.LocalDate;

public record PassengerDto(
        String firstName,
        String lastName,
        LocalDate dateOfBirth,
        String documentNumber
) {}